+++ Engine/renderer/headers/circle.hpp
@@ -4,7 +4,7 @@
 
 class Circle : public Shape {
     public:
-        Circle(Vec2<int>& center, int r, Vec2<double>& v, int m) : mCenter(center), mr(r), mv(v), mMass(m) { }
+        Circle(Vec2<double>& center, int r, Vec2<double>& v, int m) : mCenter(center), mr(r), mv(v), mMass(m) { }
         ~Circle() {}
 
         /* Render */
@@ -15,20 +15,20 @@ class Circle : public Shape {
         void resize(int factor) override;
 
         // Setters
-        void setCenter(const Vec2<int> &v) override;
-        void setXCenter(int x) override {mCenter.x = x;}
-        void setYCenter(int y) override {mCenter.y = y;}
+        void setCenter(const Vec2<double> &v) override;
+        void setXCenter(double x) override {mCenter.x = x;}
+        void setYCenter(double y) override {mCenter.y = y;}
         void setXVel(double vx) override { mv.x = vx; }
         void setYVel(double vy) override { mv.y = vy; }
 
         // Getters
-        Vec2<int> getCenter() const override;
+        Vec2<double> getCenter() const override;
         Vec2<double> getVelocity() const override;
         int getRadius() const override { return mr; }
         int getMass() const override { return mMass; }
 
     private:
-        Vec2<int> mCenter;
+        Vec2<double> mCenter;
         Vec2<double> mv;
         int mr;
         int mMass;
